♻️ 공통 fetcher함수 안에 auth() 때문에 actions.ts파일로 해당 함수 이동
