+++ util/actions.ts
@@ -1,7 +1,8 @@
 "use server";
-import { signIn } from "@/auth";
+import { auth, signIn } from "@/auth";
 import { AuthError } from "next-auth";
-import { LoginBody } from "./apiType";
+import { FetchMethod, LoginBody } from "./apiType";
+import { BASE_URL } from "./api";
 
 export async function authenticate(
   prevState: string | undefined,
@@ -27,4 +28,44 @@ export async function authenticate(
     }
     return "Unknown error occurred";
   }
+}
+
+//공통 fetcher함수(use server로 인해 해당 구역으로 이동)
+export async function fetcher(
+  endpoint: string,
+  method: FetchMethod,
+  body?: Object,
+  token?: string,
+) {
+  // 서버 측에서 next-auth에서 설정한 session값 확인
+  const session = await auth();
+  const accessToken = session?.accessToken;
+  console.log(session);
+  console.log(accessToken);
+  // 요청에 필요한 헤더 설정
+  const headers = new Headers();
+  if (accessToken) {
+    headers.append("Authorization", `Bearer ${accessToken}`);
+  }
+
+  if (body) {
+    headers.append("Content-Type", "application/json");
+  }
+
+  const options: RequestInit = {
+    method,
+    headers,
+    body: body ? JSON.stringify(body) : null,
+  };
+
+  // 원격 서버로 요청을 전송
+  const response = await fetch(`${BASE_URL}${endpoint}`, options);
+
+  if (!response.ok) {
+    const errorResponse = await response.json();
+    console.error(errorResponse.message);
+    throw new Error(errorResponse.message);
+  }
+
+  return response.json();
 }
